feat(TSP version 3) : now calculates hardest deliveries in 1.6 second

+++ application/src/main/java/fr/delivrooom/application/model/tsp/TSP1.java
@@ -8,7 +8,7 @@ import java.util.Iterator;
 public class TSP1 extends TemplateTSP {
 
     @Override
-    protected int bound(Long sommetCourant, Collection<Long> nonVus) {
+    protected float bound(Long sommetCourant, Collection<Long> nonVus) {
         return 0;
     }
